refactor(perl-lexer): split checkpoint/tests into submodules (#0000)
